psi-daemon: Do onetime write to lowmem_reserve_ratio
This config option reserves movable_zone_size / lowmem_reserve_ratio
pages from the normal zone from being used by userspace memory. This is
important in order to ensure that the kernel can continue to run while
userspace processes are stuck waiting for memory hotplug to add
additional zone movable memory.

Complications:
1)
When virtio-mem plugs in/out memory, lowmem_reserve is not updated.
It is instead updated by memory_hotplug, when adding or removing a 128Mb
block. But with virtio-mem enabled, this 128Mb block only contains 4Mb
managed pages (initially), resulting in a much smaller lowmem_reserve
than expected.

lowmem_reserve is also recalculated when userspace writes to the
lowmem_reserve_ratio sysfs file.

2)
lowmem_reserve_ratio can only be written by root; psi-deamon cannot
write to it.

Approach:
First, kernel creates an initial movable zone of size N on bootup.
Then systemd, which runs as root, writes '1' to lowmem_reserve_ratio,
as part of the process of launching psi-daemon.

This approach fails if the movable zone size becomes zero in the future.
However, this should not happen because we are setting lowmem_reserve to
be sufficiently large to prevent all userspace memory from fitting
inside the normal zone.

Experimentally, a lowmem_reserve of 512 pages was found to be sufficient
for the C2PA testcase to pass.

Change-Id: I35a3a804237fb62362d648bf4a9a49fcab587855
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>