psi_daemon: improvements to avoid the extended wakelock
1)
In the event of unplug of memory blocks, skip retrying as it can cause
unnecessary delay with wakelock held. In the below, even after
unplugging a block, it is retried with a retry_count (of 2 with delay of
10s) resulting into holding of wakelock for 20s more.

Logs:
[414.502453] pm_wake_lock: psi_daemon_lock
[424.585149] virtio_memsoc:virtio_mem_device: plugged size: 0x4800000
[424.585157] virtio_memsoc:virtio_mem_device: requested size: 0x4400000
[444.684700] pm_wake_unlock:psi_daemon_lock

This is avoided by not unplugging other memory blocks even if atleast
one block is unplugged, as this indicates that other memory blocks are
under use. This is determined by 'mem_chunks_unplugged'.

2)
In the event of unplug (by memtrack thread) followed by plug(by
mempressure event), wakelock is acquired for 10s. Worst case is
happening of these events in a loop because unplug operation is done
without considering the need of the memory block. Avoid this by trying
to acquire the wakelock after 10s delay. If suspend happens within 10s,
the memory is plugged to tvm longer than it needed.

3)
The overestimate of reclaimable memory can cause repeated plug/unplug of
memory block. Increasing the threshold by one block can make it less
likely hood of calling memtrack and acquiring the wakelock. Side effect
is that a memory block of memory may be plugged longer than necessary.
This is taken care with not going for unplug if 'count <= 1'.

All these cases can be avoided if the tvm can be brought out of suspend
by means of a timer which is not working as expected.

Change-Id: I18b73e0195b0476c743c29767e2f0aa8f1144c52
Signed-off-by: Charan Teja Kalla <quic_charante@quicinc.com>